src/ai/insight-agent/requirements-vertex.txt
flask
gunicorn
google-cloud-aiplatform
